Timer update & add old timer

+++ utils/HALDEFS/hal_defs.h
@@ -65,7 +65,7 @@
 
 #define WEEK_MS      ((uint32_t)(7 * DAY_MS))
 
-#define BITS_IN_BYTE (8)
+#define BITS_IN_BYTE ((uint8_t)8)
 
 
 #endif
